+++ NEWS.md
@@ -4,6 +4,7 @@
 * Small bugfix where crosstalk `key` and `group` was not initialised in correct FUN. 
 * Added support for button to trigger nodes and edges dynamic addition.
 * `sg_get_layout` returns nodes with coordinates rather than just the coordinates.
+* Graphs open in browser by default + warning is sent to console when using RStudio.
 
 # sigmajs 0.1.0